skill-egregore/pyproject.toml
[project]
name = "egregore-skill"
version = "2.0.0"
description = "Hive Mind Memory CLI - Advanced interface for Egregore"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests>=2.31.0",
    "sseclient-py>=1.8.0",
]

[project.scripts]
egregore = "src.main:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.6.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.mypy]
python_version = "3.13"
strict = true
